history style was fixed
