+++ frontend/src/contants/colors.ts
@@ -1,0 +1,33 @@
+export const COLORS = {
+  Beige: 'beige',
+  Black: 'black',
+  Blue: 'blue',
+  Chocolate: 'chocolate',
+  'Mocha coffee': 'mochaCoffee',
+  Pink: 'pink',
+  Pistachio: 'pistachio',
+  Yellow: 'yellow',
+  White: 'white',    
+}
+
+export const PRODUCT_HEX: { [color:string]: string } = {
+  beige: '#B19F93',
+  black: '#29211E',
+  blue: '#799EC1',
+  chocolate: '#3E2F28',
+  'mochaCoffee': '#85706C',
+  pink: '#D9BCC1',
+  pistachio: '#B3BDA2',
+  yellow: '#FFE8A3',
+  white: '#CBCBCB',
+};
+
+export const PAYMENT_STATUS_HEX: { [color:string]: string } = {
+  "awaiting payment": '#DDD0C8',  
+  "payment recieved": '#B3BDA2',
+  "shipped": '#B3BDA2',
+  "completed": '#B3BDA2',
+  "refunded": '#DDD0C8',
+  "failed": '#D9BCC1',
+  "canceled": '#D9BCC1',
+};
