+++ frontend/src/contants/endpoints.ts
@@ -1,4 +1,5 @@
 export const BASE_URL = 'http://127.0.0.1:8080/api/';
+export const MEDIA_URL = 'http://127.0.0.1:8080'
 
 export const USER = {
   TOKEN: {
