+++ frontend/src/contants/others.ts
@@ -1,0 +1,23 @@
+
+export const CATEGORIES_FILTER: { [category: string]: string } = {
+  'all': 'View All',
+ 'hoodie': 'Hoodies',
+  'tShirt': 'T-Shirts',
+  'coat': 'Coats',
+  'cropTopAndShort':'Crop Top and Shorts Sets',    
+}
+
+
+export const SIZES = {
+  S: 's',
+  M: 'm',
+  Oversize: 'oversize',
+}
+
+export const SORT_BY: { [sort:string]: string } = {
+  'newest': 'date, new to old',
+  'cheapest': 'price, low to high',
+  'mostExpensive': 'price, high to low',
+}
+
+export const ITEMS_PER_PAGE = 12;
