+++ frontend/src/pages/AccountHistoryPage/AccountHistoryPage.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react';
 import { getUserHistory } from '../../api/user';
 import { OrderResponse } from '../../types/Order';
 import { Loader } from '../../components/Loader';
-import { OrderItem } from '../../components/OrderItem/OrderItem';
+import { Order } from '../../components/Order/Order';
 
 export const AccountHistoryPage = () => {
   const [orders, setOrders] = useState<OrderResponse[]>([]);
@@ -48,7 +48,7 @@ export const AccountHistoryPage = () => {
         <ul className="AccountHistoryPage__list">
           {orders.map(order => (
             <li key={order.id}>
-              <OrderItem order={order} />
+              <Order order={order} />
             </li>
           ))}
         </ul>
