+++ frontend/src/pages/ProductDetailsPage/ProductDetailsPage.tsx
@@ -8,13 +8,13 @@ import { Product } from '../../types/Product';
 import { createSlug, getProductById } from '../../helpers/helpers';
 import { NotFoundPage } from '../NotFoundPage';
 import { getProductDetails, getProducts } from '../../api/shop';
-import { COLORS_HEX } from '../../helpers/constants';
 import { AddToCartButton } from '../../components/AddToCartButton';
 import { FavouritesContext } from '../../context/FavContext';
 import initialProducts from '../../data/products.json';
 import { BreadCrumbs } from '../../components/BreadCrumbs';
 // import productExample from '../../data/products/beige-hoodie-beige.json';
 import productExample from '../../data/products/blue-hoodie-blue.json';
+import { PRODUCT_HEX } from '../../contants/colors';
 // import productExample from '../../data/products/pink-cropTopAndShort-pink.json';
 // import productExample from '../../data/products/pistachio-tShirt-pistachio.json';
 
@@ -116,7 +116,7 @@ export const ProductDetailsPage = () => {
                 >
                   <Link
                     style={{
-                      backgroundColor: COLORS_HEX[colorValue],
+                      backgroundColor: PRODUCT_HEX[colorValue],
                     }}
                     to={`/shop/products/${category}-${colorValue}-${size.value}`}
                     className="ProductDetailsPage__colors-link"
