+++ frontend/src/styles/animations.scss
@@ -15,3 +15,14 @@
     transform: translateX(0);
   }
 }
+
+@keyframes fade-in-up {
+  0% {
+    opacity: 0;
+    transform: translateY(-120px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
