+++ frontend/src/styles/icon.scss
@@ -18,7 +18,7 @@
   }
 
   &:hover {
-    transform: scale(1.2);
+    transform: scale(1.1);
   }
 
   &--account {
